Split out <SelectOption>
